Bug Fix and Starting on Conditional Probabilites
Found a bug in the P_matrix function which was throwing off all the optimization attempts. nlm is now working and changing all the parameters.

Starting work on pseudoresiduals. Was trying to just modify code from the Zucchini and MacDonald book but I couldn't get it to work. It looks like it's rather specific to the discrete case.
I think I know a way to get pseudoresiduals from one of the definitions of conditional probability distributions in chapter 5