Data Generation, More Tweaking on Pseudoresids
Made some Data Generation function to test pseudoresiduals.

Had to rework the conditional function to the mvn.cdf function because of underflow for larger datasets.

Pseudoresiduals seem to be very close, but they are centred around -1 instead of 0. The variance ends up close to 1. I'm not sure what the problem is.